- replaced chinese text on text category buttons in lesson menu with icons/symbols - removed a warning about signed/unsigned comparison - don't render buttons for unavailable text categories on lesson menu entries - bugfix: missing initialization of text category availability bits (could lead to annoying exceptions about unavailable text categories) - disable explode button on books without lessons

+++ trunk/chinese-touch/source/text_view.cpp
@@ -429,7 +429,7 @@ ButtonAction TextView::handle_idle_cycles()
 	}
 	// render missing text lines until no characters where consumed within two consecutive iterations:
 	else if( (this->prev_size != this->char_list.size())
-				&& ((this->size() * TextView::LINE_HEIGHT + this->y_offset)
+				&& ((int)(this->size() * TextView::LINE_HEIGHT + this->y_offset)
 					<= this->text_screen.res_y) ) 
 	{
 		if( this->loading_symbol.hidden )
